Correct network id, remove dapp page, add wallet connect btn
